Rewrite badurltypes2.php.

+++ badurltypes2.php
@@ -3,39 +3,77 @@ require_once('database.inc.php');
 my_title();
 
 ini_set('max_execution_time', 180);
+$rules = array(
+	'amazon asin' =>		"url NOT ILIKE 'http://%amazon.%/%/%/%'",
+	'cover art link' =>		"(url NOT ILIKE 'http://%archive.org/%' AND url NOT ILIKE 'http://%cdbaby.%/%' AND url NOT ILIKE 'http://%jamendo.%/album/%')",
+	'IMDb' =>				"url NOT ILIKE 'http://%imdb.%/%'",
+	'musicmoz' =>			"url NOT ILIKE 'http://%musicmoz.org/%/%/%'",
+	'discogs' =>			"url NOT ILIKE 'http://%discogs.%/%'",
+	'purevolume' =>			"url NOT ILIKE 'http://www.purevolume.com/%'",
+	'myspace' =>			"url NOT ILIKE 'http://%myspace.com/%'",
+	'wikipedia' =>			"url NOT ILIKE 'http://%.wikipedia.org/%'"
+);
+
+$types = array('album', 'label', 'artist');
+
+$lttypes = array();
+foreach ($types as $type)
+{
+	$rohs = pg_query("SELECT name,linkphrase FROM lt_{$type}_url");
+	while($row = pg_fetch_assoc($rohs))
+		$lttypes[$row['name']] = $row['linkphrase'];
+}
+
+$ruledesc = array();
+
+foreach ($lttypes as $key => $lph)
+{
+	$b = isset($rules[$key]);
+	$ruledesc[] = ($b ? 'Rule' : '<span style="font-weight: bold">No rule</span>') . " for $key ('$lph')" . ($b ? ": <span style=\"font-family: monospace\">{$rules[$key]}</span>" : '') . '.';
+}
+sort($ruledesc);
 
 function gen_sql($thing)
 {
-	return "select '" . ($thing == 'album' ? 'release' : $thing) . "'::text as url_type, {$thing}.name, link0 as thing_id, link1 as urlid, linkphrase, url.url
+	global $rules;
+	$sql = "select '" . ($thing == 'album' ? 'release' : $thing) . "'::text as url_type, l_{$thing}_url.id as link_id, {$thing}.name, link0 as thing_id, link1 as urlid, linkphrase, url.url
 		from l_{$thing}_url
 		join lt_{$thing}_url on (l_{$thing}_url.link_type = lt_{$thing}_url.id)
 		join url on (url.id = link1)
 		join {$thing} on ({$thing}.id = link0)
-		WHERE
-		(lt_{$thing}_url.name = 'amazon asin' AND url NOT ILIKE 'http://%amazon.%/%/%/%') OR
-		(lt_{$thing}_url.name = 'cover art link' AND (url NOT ILIKE 'http://%archive.org/%' AND url NOT ILIKE 'http://%cdbaby.%/%' AND url NOT ILIKE 'http://%jamendo.%/{$thing}/%')) OR
-		(lt_{$thing}_url.name = 'IMDb' AND url NOT ILIKE 'http://%imdb.%/%') OR
-		(lt_{$thing}_url.name = 'musicmoz' AND url NOT ILIKE 'http://%musicmoz.org/%/%/%') OR
-		(lt_{$thing}_url.name = 'discogs' AND url NOT ILIKE 'http://%discogs.%/%') OR
-		(lt_{$thing}_url.name = 'wikipedia' AND url NOT ILIKE 'http://%.wikipedia.org/%');";
+		WHERE ";
+
+	foreach ($rules as $rule => $data)
+		$sql .= "(lt_{$thing}_url.name = '{$rule}' AND $data) OR \n";
+
+	$sql .= "FALSE;"; // LAZY
+	return $sql;
 }
 
-$sql = 'begin;
-drop table if exists bad_url_types;
-create table bad_url_types as ' . gen_sql('album');
+if (isset($_GET{'regenerate'}))
+{
+	ignore_user_abort(true);
+	ini_set('max_execution_time', 180);
 
-foreach (array('label', 'artist') as $thing)
-	$sql .= 'insert into bad_url_types ' . gen_sql($thing) . "\n\n";
+	$sql = 'begin;
+	drop table if exists bad_url_types;
+	create table bad_url_types as ' . gen_sql(array_pop($types));
 
-$sql .= 'commit;';
+	foreach ($types as $thing)
+		$sql .= 'insert into bad_url_types ' . gen_sql($thing) . "\n\n";
 
-pg_query($sql) or die (pg_last_error());
+	$sql .= 'commit;';
 
-$res = pg_query("select * from bad_url_types
+	pg_query($sql) or die (pg_last_error());
+	die('Done. <a href="' . $_SERVER['SCRIPT_NAME'] . '">Go back</a>.');
+}
+
+$res = pg_query("select * from bad_url_types " .
+	(isset($_GET{'catcatlovesthepony'}) ? "WHERE url_type = 'release' " : '') . "
 	ORDER BY linkphrase ASC,url_type ASC,url ASC
 ");
 
-echo '<p>' . pg_num_rows($res) . ' found.</p>';
+echo '<p>' . pg_num_rows($res) . ' found. <a href="?regenerate">Regenerate cache (&lt; one minute).</a></p>';
 
 ?><table><tr><th>Thing</th><th>URL</th></tr><?
 
@@ -43,11 +81,18 @@ while ($row = pg_fetch_assoc($res))
 {
 	if (isset($last) && $row['linkphrase'] != $last)
 		echo '<tr><td colspan="2"><hr/></td></tr>';
-	echo '<tr><td>' . $row['url_type'] . ': <a href="http://musicbrainz.org/show/' . $row['url_type'] . '?' . $row['url_type'] . 'id=' . $row['thing_id'] . '">' . $row['name'] . '</a> ' . ($last = $row['linkphrase']) . '</td><td>[ <a href="http://musicbrainz.org/show/url/?urlid=' . $row['urlid'] . '">show</a> ] [ <a href="http://musicbrainz.org/edit/url/edit.html?urlid=' . $row['urlid'] . '">edit</a> ] <a href="' . $row['url'] . '">' . $row['url'] . '</a></td>' . "\n";
+	echo '<tr><td>' . $row['url_type'] . ': <a href="http://musicbrainz.org/show/' . $row['url_type'] . '?' . $row['url_type'] . 'id=' . $row['thing_id'] . '">' . $row['name'] . '</a> ' .
+		($last = $row['linkphrase']) . '</td><td>' .
+		'[ <a href="http://musicbrainz.org/show/url/?urlid=' . $row['urlid'] . '">show</a> ] ' .
+		'[ <a href="http://musicbrainz.org/edit/url/edit.html?urlid=' . $row['urlid'] . '">edit</a> ] ' .
+		'[ <a href="http://musicbrainz.org/edit/relationship/remove.html?type=' . ($row['url_type'] == 'release' ? 'album' : $row['url_type']) . '-url&id=' . $row['link_id'] . '">remove</a> ] ' .
+		'<a href="' . $row['url'] . '">' . $row['url'] . '</a></td>' . "\n";
 }
 
 ?>
 </table></tr>
-<?echo "<p>Generated in " . (time()-$start) . " seconds.</p>";?>
+<?
+echo "<ul><li>" . implode('</li><li>', $ruledesc) . '</li></ul>';
+echo "<p>Generated in " . (time()-$start) . " seconds.</p>";?>
 </body>
 </html>
